chore: move Selenium test files to streamline project structure
